fix controller / ssgen

+++ pkg/infrastructure/backend/k8simpl/helper.go
@@ -78,15 +78,14 @@ func syncResources[T apiResource](ctx context.Context, cluster *discovery.Cluste
 		if err != nil {
 			return err
 		}
-		// For StatefulSets, delete the resource before applying again - StatefulSet has many immutable fields
-		// cf. updates to statefulset spec for fields other than 'replicas', 'ordinals', 'template', 'updateStrategy', 'revisionHistoryLimit', 'persistentVolumeClaimRetentionPolicy' and 'minReadySeconds' are forbidden
-		if replace {
-			err = s.Delete(ctx, rc.GetName(), metav1.DeleteOptions{PropagationPolicy: lo.ToPtr(metav1.DeletePropagationForeground)})
-			if err != nil {
-				return err
-			}
-		}
 		_, err = s.Patch(ctx, rc.GetName(), types.ApplyPatchType, b, metav1.PatchOptions{Force: lo.ToPtr(true), FieldManager: fieldManager})
+		// For StatefulSets, delete the resource before applying again - StatefulSet has many immutable fields
+		// Example: StatefulSet.apps "nsapp-add177a080c4c78936e192" is invalid: spec: Forbidden: updates to statefulset spec for fields other than 'replicas', 'ordinals', 'template', 'updateStrategy', 'revisionHistoryLimit', 'persistentVolumeClaimRetentionPolicy' and 'minReadySeconds' are forbidden
+		if replace && err != nil {
+			_ = s.Delete(ctx, rc.GetName(), metav1.DeleteOptions{PropagationPolicy: lo.ToPtr(metav1.DeletePropagationForeground)})
+			// Try again
+			_, err = s.Patch(ctx, rc.GetName(), types.ApplyPatchType, b, metav1.PatchOptions{Force: lo.ToPtr(true), FieldManager: fieldManager})
+		}
 		if err != nil {
 			return err
 		}
